settlements

+++ src/app/api/group/route.ts
@@ -4,6 +4,8 @@ import {fetchWithAuth} from "@/lib/api/fetch-with-auth";
 
 const BASE_URL = process.env.BASE_URL;
 const GROUP = process.env.GROUP;
+const UPDATE_GROUP = process.env.UPDATE_GROUP;
+const DELETE_GROUP = process.env.DELETE_GROUP;
 
 export async function POST(request: NextRequest) {
     try {
@@ -38,10 +40,18 @@ export async function POST(request: NextRequest) {
 
 export async function PUT(request: NextRequest) {
     try {
-        const {id, name, color}: GroupCreate & { id: string } = await request.json();
-        
+        const groupId = request.nextUrl.searchParams.get('groupId');
+        if (!groupId) {
+            return NextResponse.json(
+                {success: false, message: 'Brak id grupy w zapytaniu'},
+                {status: 400}
+            );
+        }
+        const {name, color}: GroupCreate = await request.json();
+
+        const endpoint = UPDATE_GROUP?.replace('{groupId}', groupId);
         const cookieHeader = request.headers.get('cookie') ?? '';
-        const response = await fetchWithAuth(`${BASE_URL}${GROUP}/${id}`, {
+        const response = await fetchWithAuth(`${BASE_URL}${endpoint}`, {
             method: 'PUT',
             headers: {
                 'Cookie': cookieHeader,
@@ -53,10 +63,7 @@ export async function PUT(request: NextRequest) {
             }),
             credentials: 'include',
         });
-
         const data = await response.json();
-
-
         return NextResponse.json(data, {status: response.status});
     } catch (error) {
         console.error('Group update API error:', error);
@@ -69,9 +76,16 @@ export async function PUT(request: NextRequest) {
 
 export async function DELETE(request: NextRequest) {
     try {
-        const {id} = await request.json();
+        const groupId = request.nextUrl.searchParams.get('groupId');
+        if (!groupId) {
+            return NextResponse.json(
+                {success: false, message: 'Brak id grupy w zapytaniu'},
+                {status: 400}
+            );
+        }
+        const endpoint = DELETE_GROUP?.replace('{groupId}', groupId);
         const cookieHeader = request.headers.get('cookie') ?? '';
-        const response = await fetchWithAuth(`${BASE_URL}${GROUP}/${id}`, {
+        const response = await fetchWithAuth(`${BASE_URL}${endpoint}`, {
             method: 'DELETE',
             headers: {
                 'Cookie': cookieHeader,
@@ -81,8 +95,6 @@ export async function DELETE(request: NextRequest) {
         });
 
         const data = await response.json();
-
-
         return NextResponse.json(data, {status: response.status});
     } catch (error) {
         console.error('Group deletion API error:', error);
